examples to separate folder, still as tests
